move utility classes to flink-fs-common
