+++ flink-filesystems/flink-fs-common/src/main/java/org/apache/flink/fs/common/util/RefCountedBufferingFileStream.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.fs.s3.common.utils;
+package org.apache.flink.fs.common.util;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.annotation.VisibleForTesting;
