+++ flink-filesystems/flink-s3-fs-base/src/main/java/org/apache/flink/fs/s3/common/writer/S3RecoverableFsDataOutputStream.java
@@ -21,9 +21,9 @@ package org.apache.flink.fs.s3.common.writer;
 import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.core.fs.RecoverableFsDataOutputStream;
 import org.apache.flink.core.fs.RecoverableWriter;
-import org.apache.flink.fs.s3.common.utils.RefCountedBufferingFileStream;
-import org.apache.flink.fs.s3.common.utils.RefCountedFSOutputStream;
-import org.apache.flink.fs.s3.common.utils.RefCountedFile;
+import org.apache.flink.fs.common.util.RefCountedBufferingFileStream;
+import org.apache.flink.fs.common.util.RefCountedFSOutputStream;
+import org.apache.flink.fs.common.util.RefCountedFile;
 import org.apache.flink.util.function.FunctionWithException;
 
 import org.apache.commons.io.IOUtils;
