+++ flink-filesystems/pom.xml
@@ -47,6 +47,7 @@ under the License.
 		<module>flink-s3-fs-presto</module>
 		<module>flink-swift-fs-hadoop</module>
 		<module>flink-oss-fs-hadoop</module>
+		<module>flink-fs-common</module>
 	</modules>
 
 	<!-- Common dependency setup for all filesystems -->
